Fix the style in .child on line 31
So the style doesn't work on .child because the
child is inside .text and inside it there is a div
with two classes text and child class. The style
of .text is more followed because it has two
classes so the style doesn't work on .child so
what I did was I used a descendant combinator
since its parent element is .text class.