Set the width of .footer to 100%
That's why I set the .footer width to 100%,
because when using the position property,
justify-content: space-between wasn’t working.
The width of the .footer seemed to shrink, so the
items inside were still close together even though
space-between was applied. To make space-between
work properly, I set its width to 100%.